+++ CrazyBandit/Modules/CrazyBandit.Engine/WinningsCalculator.cs
@@ -13,22 +13,22 @@ namespace CrazyBandit.Engine
         /// <summary>
         /// Wygrana za linie wygrywającą o danym symbolu
         /// </summary>
-        private readonly Dictionary<int, float> m_winnings;
+        private readonly Dictionary<int, double> m_winnings;
 
         /// <summary>
         /// Konfiguruje mnożnik wygranych za linie danego symbolu.
         /// </summary>
         /// <param name="winnings"></param>
-        public WinningsCalculator(Dictionary<int, float> winnings)
+        public WinningsCalculator(Dictionary<int, double> winnings)
         {
             Ensure.ParamNotNull(winnings, nameof(winnings));
             m_winnings = winnings;
         }
 
         /// <inheritdoc />
-        public float Calculate(int bet, int symbol)
+        public double Calculate(int bet, int symbol)
         {
-            float winningValue = m_winnings[symbol];
+            double winningValue = m_winnings[symbol];
             return bet * winningValue;            
         }
     }
